fixed package typo and added aerogear-parent pom
