+++ pom.xml
@@ -5,14 +5,11 @@
   <name>Java Microsoft Push Notification Service Library</name>
   <version>0.0.5</version>
 
-    <!-- Inherit the Sonatype OSS deployment configuration -->
-    <!--
     <parent>
-        <groupId>org.sonatype.oss</groupId>
-        <artifactId>oss-parent</artifactId>
-        <version>5</version>
+        <groupId>org.jboss.aerogear</groupId>
+        <artifactId>aerogear-parent</artifactId>
+        <version>0.2.9</version>
     </parent>
-    -->
 
     <scm>
         <connection>scm:git:git://github.com/aerogear/java-mpns.git</connection>
@@ -20,14 +17,6 @@
         <url>http://github.com/aerogear/java-mpns</url>
     </scm>
 
-   <licenses>
-      <license>
-         <name>The Apache Software License, Version 2.0</name>
-         <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-         <distribution>repo</distribution>
-      </license>
-   </licenses>
-    
     <properties>
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
@@ -189,59 +178,4 @@
     <description>
         Library to send Microsoft Push Notification.
     </description>
-
-    
-    <reporting>
-      <plugins>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-javadoc-plugin</artifactId>
-          <version>2.7</version>
-          <configuration>
-              <excludePackageNames>com.notnoop.mpns.internal</excludePackageNames>
-          </configuration>
-        </plugin>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-jxr-plugin</artifactId>
-          <version>2.2</version>
-        </plugin>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-surefire-plugin</artifactId>
-          <version>2.7.1</version>
-        </plugin>
-        <plugin>
-          <groupId>com.atlassian.maven.plugins</groupId>
-          <artifactId>maven-clover2-plugin</artifactId>
-          <version>2.3.1</version>
-        </plugin>
-
-        <!--
-          <plugin>
-            <groupId>org.codehaus.mojo</groupId>
-            <artifactId>findbugs-maven-plugin</artifactId>
-            <version>2.0.1</version>
-          </plugin>
-          -->
-
-      </plugins>
-    </reporting>
-
-    <issueManagement>
-        <system>Github</system>
-        <url>http://github.com/notnoop/java-mpns/issues</url>
-    </issueManagement>
-
-    <developers>
-        <developer>
-            <id>notnoop</id>
-            <name>Mahmood Ali</name>
-        </developer>
-        <developer>
-            <id>edewit</id>
-            <name>Erik Jan de Wit</name>
-        </developer>        
-    </developers>
-
 </project>
