--- src/main/java/org/jboss/aerogeaer/windows/mpns/MPNS.java
@@ -1,44 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright Red Hat, Inc., and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.aerogeaer.windows.mpns;
-
-/**
- * The main class to interact with the MPNS Service.
- *
- * Provides an interface to create the {@link MpnsNotificationBuilder} and
- * {@link MpnsNotification} payload.
- *
- */
-public class MPNS {
-
-    private MPNS() { throw new AssertionError("Uninstantiable class"); }
-
-    /**
-     * Returns a new Payload builder
-     */
-    public static MpnsNotificationBuilder newNotification() {
-        return new MpnsNotificationBuilder();
-    }
-
-    /**
-     * Returns a new APNS Service for sending iPhone notifications
-     */
-    public static MpnsServiceBuilder newService() {
-        return new MpnsServiceBuilder();
-    }
-}
