--- src/main/java/org/jboss/aerogeaer/windows/mpns/MpnsNotification.java
@@ -1,42 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright Red Hat, Inc., and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.aerogeaer.windows.mpns;
-
-import java.util.List;
-import java.util.Map.Entry;
-
-/**
- * Represents an MPNS notification to be sent to Microsoft service.
- */
-public interface MpnsNotification {
-
-    /**
-     * Returns the binary representation of the payload.
-     *
-     */
-    public byte[] getRequestBody();
-
-    /**
-     * Returns the identifier of the current message.  The
-     * identifier is an application generated identifier.
-     *
-     * @return the notification identifier
-     */
-    public List<? extends Entry<String, String>> getHttpHeaders();
-    
-}
