--- src/main/java/org/jboss/aerogeaer/windows/mpns/notifications/RawNotification.java
@@ -1,96 +1,0 @@
-/*
- * Copyright 2011, Mahmood Ali.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *   * Redistributions in binary form must reproduce the above
- *     copyright notice, this list of conditions and the following disclaimer
- *     in the documentation and/or other materials provided with the
- *     distribution.
- *   * Neither the name of Mahmood Ali. nor the names of its
- *     contributors may be used to endorse or promote products derived from
- *     this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.jboss.aerogeaer.windows.mpns.notifications;
-
-import org.jboss.aerogeaer.windows.mpns.DeliveryClass;
-import org.jboss.aerogeaer.windows.mpns.MpnsNotification;
-import org.jboss.aerogeaer.windows.mpns.internal.Utilities;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map.Entry;
-
-public class RawNotification implements MpnsNotification {
-    private final byte[] body;
-
-    private final List<? extends Entry<String, String>> headers;
-
-    public RawNotification(byte[] body,
-            List<? extends Entry<String, String>> headers) {
-        this.body = body;
-
-        this.headers = headers;
-    }
-
-    public byte[] getRequestBody() {
-        return body;
-    }
-
-    public List<? extends Entry<String, String>> getHttpHeaders() {
-        return Collections.unmodifiableList(this.headers);
-    }
-
-    public static class Builder extends AbstractNotificationBuilder<Builder, RawNotification> {
-        private byte[] body;
-
-        public Builder() {
-            super();
-        }
-
-        public Builder body(String body) {
-            this.body = Utilities.toUTF8(body);
-            return this;
-        }
-
-        public Builder body(byte[] body) {
-            byte[] copy = new byte[body.length];
-            System.arraycopy(body, 0, copy, 0, body.length);
-            this.body = copy;
-            return this;
-        }
-
-        @Override
-        protected int deliveryValueOf(DeliveryClass delivery) {
-            switch (delivery) {
-            case IMMEDIATELY:   return 3;
-            case WITHIN_450:    return 13;
-            case WITHIN_900:    return 23;
-            default:
-                throw new AssertionError("Unknown Value: " + delivery);
-            }
-        }
-
-        @Override
-        public RawNotification build() {
-            return new RawNotification(body, headers);
-        }
-    }
-}
