+++ src/main/java/org/jboss/aerogear/windows/mpns/DeliveryClass.java
@@ -1,0 +1,45 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright Red Hat, Inc., and individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.aerogear.windows.mpns;
+
+/**
+ * Determines the batch interval indicating when the push notification
+ * should be sent to the application.
+ *
+ * The default value, for a new notification is {@code IMMEDIATELY}.
+ *
+ */
+public enum DeliveryClass {
+
+    /**
+     * Indicates that the push notification should be delivered immediately
+     */
+    IMMEDIATELY,
+
+    /**
+     * Indicates that the push notification should be delivered within 450
+     * seconds (7.5 minutes)
+     */
+    WITHIN_450,
+
+    /**
+     * Indicates that the push notification should be delivered within 900
+     * seconds (15 minutes)
+     */
+    WITHIN_900
+}
