+++ src/main/java/org/jboss/aerogear/windows/mpns/internal/Pair.java
@@ -1,0 +1,46 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright Red Hat, Inc., and individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.aerogear.windows.mpns.internal;
+
+import java.util.Map;
+
+public class Pair<K, V> implements Map.Entry<K, V> {
+    public final K key;
+    public final V value;
+
+    public Pair(K k, V v) {
+        this.key = k;
+        this.value = v;
+    }
+
+    public K getKey() {
+        return key;
+    }
+
+    public V getValue() {
+        return value;
+    }
+
+    public V setValue(V value) {
+        throw new UnsupportedOperationException("Pair is immutable!");
+    }
+
+    public static <K, V> Pair<K, V> of(K k, V v) {
+        return new Pair<K, V>(k, v);
+    }
+}
