--- src/test/java/org/jboss/aerogeaer/windows/mpns/MPNSTest.java
@@ -1,45 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright Red Hat, Inc., and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.aerogeaer.windows.mpns;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-import static org.hamcrest.CoreMatchers.*;
-
-/**
- * Silly Tests
- */
-public class MPNSTest {
-
-    @Test
-    public void testInstances() {
-        assertThat(MPNS.newNotification(), is(MpnsNotificationBuilder.class));
-        assertThat(MPNS.newService(), is(MpnsServiceBuilder.class));
-    }
-
-    @Test
-    public void notificationShouldGetNewInstances() {
-        assertNotSame(MPNS.newNotification(), MPNS.newNotification());
-    }
-
-    @Test
-    public void newServiceGetNewInstances() {
-        assertNotSame(MPNS.newService(), MPNS.newService());
-    }
-}
