--- src/test/java/org/jboss/aerogeaer/windows/mpns/notifications/TileNotificationTest.java
@@ -1,37 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright Red Hat, Inc., and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.aerogeaer.windows.mpns.notifications;
-
-import org.jboss.aerogeaer.windows.mpns.MPNS;
-import org.junit.Test;
-
-import static org.jboss.aerogeaer.windows.mpns.util.matches.IsValidXml.isValidXml;
-import static org.junit.Assert.assertThat;
-
-public class TileNotificationTest {
-
-    @Test
-    public void generateValidXml() {
-        TileNotification tile = MPNS.newNotification().tile()
-            .count(1).title("Sample").backBackgroundImage("http://sample.com/image.png")
-            .backTitle("Random \"asdf<>;'a")
-            .build();
-
-        assertThat(tile.getRequestBody(), isValidXml());
-    }
-}
