+++ src/test/java/org/jboss/aerogear/windows/mpns/internal/UtilitiesTest.java
@@ -1,0 +1,37 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright Red Hat, Inc., and individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.aerogear.windows.mpns.internal;
+
+import org.junit.Test;
+
+import static org.jboss.aerogear.windows.mpns.internal.Utilities.ifNonNull;
+import static org.junit.Assert.*;
+import static org.hamcrest.CoreMatchers.*;
+
+public class UtilitiesTest {
+
+    @Test
+    public void ifNonNullHandlesNull() {
+        assertThat(ifNonNull(null, "asdf"), is(""));
+    }
+
+    @Test
+    public void isNonNullHandlesNonNull() {
+        assertThat(ifNonNull("A", "asdf" + "fdsa"), is("asdffdsa"));
+    }
+}
